+++ .eslintrc.cjs
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: ['alloy', 'alloy/typescript', 'prettier'],
-  env: {
-    node: true,
-  },
-};